requirements.txt
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
filelock==3.7.1
gdown==4.4.0
google-api-core==2.8.1
google-api-python-client==2.50.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
google-auth==2.7.0
googleapis-common-protos==1.56.2
httplib2==0.20.4
idna==3.3
numpy==1.22.4; platform_machine != 'aarch64' and platform_machine != 'arm64' and python_version < '3.10'
oauthlib==3.2.0
pandas==1.4.2
protobuf==3.20.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pymupdf==1.19.6
pyparsing==3.0.9; python_version > '3.0'
pypdf2==2.1.0
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests-oauthlib==1.3.1
requests==2.28.0
rsa==4.8; python_version >= '3.6'
six==1.16.0
soupsieve==2.3.2.post1
tqdm==4.64.0
typing-extensions==4.2.0; python_version < '3.10'
uritemplate==4.1.1
urllib3==1.26.9
